[CARBONDATA-3889] Cleanup code in carbondata-core module

+++ core/src/main/java/org/apache/carbondata/core/datastore/page/UnsafeVarLengthColumnPage.java
@@ -39,7 +39,7 @@ public class UnsafeVarLengthColumnPage extends UnsafeVarLengthColumnPageBase {
     } else {
       capacity = (int) (pageSize * DEFAULT_ROW_SIZE * FACTOR);
     }
-    memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, (long) (capacity));
+    memoryBlock = UnsafeMemoryManager.allocateMemoryWithRetry(taskId, capacity);
     baseAddress = memoryBlock.getBaseObject();
     baseOffset = memoryBlock.getBaseOffset();
   }
